Cypress/E2E: Phase 1 Move specs to enterprise folder
