+++ e2e/cypress/integration/enterprise/accessibility/accessibility_modals_dialogs_spec.js
@@ -8,9 +8,9 @@
 // ***************************************************************
 
 // Stage: @prod
-// Group: @accessibility
+// Group: @enterprise @accessibility
 
-import * as TIMEOUTS from '../../fixtures/timeouts';
+import * as TIMEOUTS from '../../../fixtures/timeouts';
 
 let selectedRowText;
