+++ e2e/cypress/integration/enterprise/cloud/onboarding/finish_tutorial_spec.js
@@ -8,10 +8,10 @@
 // - Use element ID when selecting an element. Create one if none.
 // ***************************************************************
 
-// Group: @onboarding
+// Group: @enterprise @onboarding
 
-import * as TIMEOUTS from '../../fixtures/timeouts';
-import {generateRandomUser} from '../../support/api/user';
+import * as TIMEOUTS from '../../../../fixtures/timeouts';
+import {generateRandomUser} from '../../../../support/api/user';
 
 describe('Onboarding', () => {
     let testTeam;
