+++ e2e/cypress/integration/enterprise/profile_popover/profile_popover_spec_user_a_b_spec.js
@@ -8,7 +8,7 @@
 // ***************************************************************
 
 // Stage: @prod
-// Group: @profile_popover
+// Group: @enterprise @profile_popover
 
 describe('Profile popover User A & B', () => {
     let testTeam;
