+++ e2e/cypress/integration/enterprise/system_console/group_configuration_spec.js
@@ -8,9 +8,9 @@
 // ***************************************************************
 
 // Stage: @prod
-// Group: @system_console
+// Group: @enterprise @system_console
 
-import * as TIMEOUTS from '../../fixtures/timeouts';
+import * as TIMEOUTS from '../../../fixtures/timeouts';
 
 describe('group configuration', () => {
     let groupID;
