+++ e2e/cypress/integration/enterprise/system_console/search_box_spec.js
@@ -1,0 +1,47 @@
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+// ***************************************************************
+// - [#] indicates a test step (e.g. # Go to a page)
+// - [*] indicates an assertion (e.g. * Check the title)
+// - Use element ID when selecting an element. Create one if none.
+// ***************************************************************
+
+// Group: @enterprise @system_console
+
+import * as TIMEOUTS from '../../../fixtures/timeouts';
+
+// # Goes to the System Scheme page as System Admin
+const goToAdminConsole = () => {
+    cy.apiAdminLogin();
+    cy.visit('/admin_console');
+};
+
+describe('System console', () => {
+    before(() => {
+        // * Check if server has license
+        cy.apiRequireLicense();
+    });
+
+    it('MM-T897_1 - Focus should be in System Console search box on opening System Console or refreshing pages in System Console', () => {
+        const pageIds = ['reporting\\/system_analytics', 'reporting\\/team_statistics', 'reporting\\/server_logs', 'user_management\\/users', 'user_management\\/teams'];
+        goToAdminConsole();
+
+        // * Assert the ID of the element is the ID of admin sidebar filter
+        cy.focused().should('have.id', 'adminSidebarFilter');
+        cy.wait(TIMEOUTS.ONE_SEC);
+
+        pageIds.forEach((id) => {
+            // # Go to another page
+            cy.get(`#${id}`).click();
+
+            // * Ensure focus is lost
+            cy.focused().should('not.have.id', 'adminSidebarFilter');
+
+            // * Reload and ensure the focus is back on the search component
+            cy.reload();
+            cy.focused().should('have.id', 'adminSidebarFilter');
+            cy.wait(TIMEOUTS.ONE_SEC);
+        });
+    });
+});
