+++ e2e/cypress/integration/search_autocomplete/users_spec.js
@@ -18,9 +18,9 @@ describe('Autocomplete without Elasticsearch - Users', () => {
     let testTeam;
 
     before(() => {
-        // * Check if server has license for Elasticsearch
-        cy.apiRequireLicenseForFeature('Elasticsearch');
-
+        // # Remove licese
+        cy.apiDeleteLicense();
+        
         // # Create new team for tests
         cy.apiCreateTeam(`search-${timestamp}`, `search-${timestamp}`).then(({team}) => {
             testTeam = team;
